Controller refactored, border button color & hover changed
