+++ Library-webapp/src/main/java/org/alain/library/webapp/web/ReservationController.java
@@ -1,0 +1,4 @@
+package org.alain.library.webapp.web;
+
+public class ReservationController {
+}
